Podzielono na klasy (nic nie dziala)
